requirements.txt
# MCP and Elasticsearch dependencies
mcp>=1.0.0
elasticsearch
elasticsearch[async]

# Environment and configuration
python-dotenv>=1.0.0
requests

# Web framework for UI
flask>=3.0.0
flask-cors>=4.0.0

# Utility libraries
asyncio-mqtt>=0.16.0
aiofiles>=23.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: For advanced logging and monitoring
structlog>=23.0.0
prometheus-client>=0.19.0
